move tests to `runtests.jl`
